c30403.07 I renamed server3/s32 to s30 since it is the original backup that works
